garner/maintainer/requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
google-auth==1.30.0
idna==2.10
kubernetes==17.17.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
urllib3==1.26.4
websocket-client==1.0.0
